requirements.txt
adam_atan2_pytorch==0.1.18
beartype==0.21.0
Bio==1.8.0
biopython==1.85
cachetools==5.5.2
click==8.2.1
CoLT5_attention==0.11.1
deepspeed==0.16.8
einops==0.8.1
ema_pytorch==0.7.7
environs==14.2.0
fair_esm==2.0.0
frame_averaging_pytorch==0.1.2
gemmi==0.7.1
gradio==5.33.1
gradio_molecule3d==0.0.7
huggingface_hub==0.31.4
jaxtyping==0.3.2
joblib==1.5.0
liger_kernel==0.5.9
lion_pytorch==0.2.3
loguru==0.7.3
multimolecule==0.0.7
numpy==2.3.0
packaging==25.0
pdbeccdutils==0.8.5
polars==1.30.0
pydantic==2.11.5
PyYAML==6.0.2
PyYAML==6.0.2
rdkit==2025.3.2
scipy==1.15.3
taylor_series_linear_attention==0.1.12
torchinfo==1.8.0
torchmetrics==1.7.1
tqdm==4.67.1
transformers==4.50.0
triton==3.3.0
wandb==0.20.1
wrapt_timeout_decorator==1.5.1
sh==2.2.2
rootutils==1.0.7
lightning==2.5.1.post0